+++ botx/testing/botx_mock/wsgi/application.py
@@ -11,6 +11,7 @@ from botx.testing.botx_mock.wsgi.routes import (
     chats,
     command,
     events,
+    files,
     notification,
     notifications,
     smartapps,
@@ -45,6 +46,9 @@ _ENDPOINTS: Tuple[Callable[..., Any], ...] = (
     users.get_by_login,
     # notifications
     notifications.post_internal_bot_notification,
+    # files
+    files.upload_file,
+    files.download_file,
     # smartapps
     smartapps.post_smartapp_event,
     smartapps.post_smartapp_notification,